requirements.txt
aiohttp
pyjwt
qrcode
pillow